Move SelectableHits out of ProzesskopieForm
