+++ Kitodo/src/main/java/de/sub/goobi/forms/ProzesskopieForm.java
@@ -45,6 +45,7 @@ import org.kitodo.data.database.beans.Process;
 import org.kitodo.data.database.beans.Project;
 import org.kitodo.data.database.beans.User;
 import org.kitodo.data.database.exceptions.DAOException;
+import org.kitodo.helper.SelectableHit;
 import org.kitodo.services.ServiceManager;
 
 import ugh.dl.Fileformat;
@@ -58,104 +59,6 @@ public class ProzesskopieForm implements Serializable {
     private static final Logger logger = LogManager.getLogger(ProzesskopieForm.class);
     private static final long serialVersionUID = -4512865679353743L;
     private transient ServiceManager serviceManager = new ServiceManager();
-
-    /**
-     * The class SelectableHit represents a hit on the hit list that shows up if
-     * a catalogue search yielded more than one result. We need an inner class
-     * for this because Faces is striclty object oriented and the always
-     * argument-less actions can only be executed relatively to the list entry
-     * in question this way if they are concerning elements that are rendered by
-     * iterating along a list.
-     *
-     * @author Matthias Ronge &lt;matthias.ronge@zeutschel.de&gt;
-     */
-    public class SelectableHit {
-        /**
-         * The field hit holds the hit to be rendered as a list entry.
-         */
-        private final Hit hit;
-
-        /**
-         * The field error holds an error message to be rendered as a list entry
-         * in case that retrieving the hit failed within the plug-in used for
-         * catalogue access.
-         */
-        private final String error;
-
-        /**
-         * Selectable hit constructor. Creates a new SelectableHit object with a
-         * hit to show.
-         *
-         * @param hit
-         *            Hit to show
-         */
-        public SelectableHit(Hit hit) {
-            this.hit = hit;
-            error = null;
-        }
-
-        /**
-         * Selectable hit constructor. Creates a new SelectableHit object with
-         * an error message to show.
-         *
-         * @param error
-         *            error message
-         */
-        public SelectableHit(String error) {
-            hit = null;
-            this.error = error;
-        }
-
-        /**
-         * The function getBibliographicCitation() returns a summary of this hit
-         * in bibliographic citation style as HTML as read-only property
-         * “bibliographicCitation”.
-         *
-         * @return a summary of this hit in bibliographic citation style as HTML
-         */
-        public String getBibliographicCitation() {
-            return hit.getBibliographicCitation();
-        }
-
-        /**
-         * The function getErrorMessage() returns an error if that had occurred
-         * when trying to retrieve that hit from the catalogue as read-only
-         * property “errorMessage”.
-         *
-         * @return an error message to be rendered as a list entry
-         */
-        public String getErrorMessage() {
-            return error;
-        }
-
-        /**
-         * The function isError() returns whether an error occurred when trying
-         * to retrieve that hit from the catalogue as read-only property
-         * “error”.
-         *
-         * @return whether an error occurred when retrieving that hit
-         */
-        public boolean isError() {
-            return hit == null;
-        }
-
-        /**
-         * The function selectClick() is called if the user clicks on a
-         * catalogue hit summary in order to import it into Production.
-         *
-         * @return always "", indicating to Faces to stay on that page
-         */
-        public String selectClick() {
-            try {
-                importHit(hit);
-            } catch (Exception e) {
-                Helper.setFehlerMeldung("Error on reading opac ", e);
-            } finally {
-                hitlistPage = -1;
-            }
-            return null;
-        }
-    }
 
     /**
      * The constant DEFAULT_HITLIST_PAGE_SIZE holds the fallback number of hits
@@ -376,7 +279,7 @@ public class ProzesskopieForm implements Serializable {
      * @param hit
      *            Hit to load
      */
-    protected void importHit(Hit hit) throws PreferencesException {
+    public void importHit(Hit hit) throws PreferencesException {
         rdf = hit.getFileformat();
         docType = hit.getDocType();
         copyProcess.fillFieldsFromMetadataFile();
@@ -896,5 +799,13 @@ public class ProzesskopieForm implements Serializable {
      */
     public Fileformat getFileformat() {
         return rdf;
+    }
+
+    public long getHitlistPage() {
+        return hitlistPage;
+    }
+
+    public void setHitlistPage(long hitlistPage) {
+        this.hitlistPage = hitlistPage;
     }
 }
